Implemented deleting, finishing, editing & updating of tasks.

+++ resources/assets/js/rest.js
@@ -1,6 +1,7 @@
-$(".datepicker").datepicker({
-    format: "yyyy-mm-dd",
-    autoclose: true
+$(".datepicker").datetimepicker({
+    minDate: Date.now(),
+    format: "YYYY-MM-DD HH:mm:ss",
+    sideBySide: true
 });
 
 $.ajaxSetup({
@@ -11,15 +12,47 @@ $.ajaxSetup({
 var baseUrl = $("meta[name='base-url']").attr("content");
 
 //<a data-taskid="3">Delete</a>
-$(".task-delete").click(function () {
+$("#tasks").on("click", ".task-delete", function () {
+    if (!confirm("Do you really want to delete this task?")) {
+        return true;
+    }
     $.ajax({
         method: "DELETE",
         url: baseUrl + "/task/" + $(this).data("taskid") + "/delete"
-    })
-    .done(function () {
-        //Refresh ui here
-    })
-    .error(function (msg) {
-        console.error("Error:", msg.responseText);
+    }).done(function (response) {
+        $("#tasks").html(response);
+    }).error(function (response) {
+        console.error("Error:", response.responseText);
     });
+}).on("change", ".task-check", function () {
+    $(this).parent().siblings(".task-description").children("p").toggleClass("done");
+    $.ajax({
+        method: "PATCH",
+        url: baseUrl + "/task/" + $(this).data("taskid") + "/check"
+    }).done(function () {
+        // Success!
+    }).error(function (response) {
+        console.error("Error:", response.responseText);
+    });
+}).on("keydown", ".task-description > p", function (e) {
+
+    // On enter click leave the edit area and update the task.
+    if (e.which == 13) {
+        e.preventDefault();
+        $(this).blur();
+
+        var description = $(this).text();
+
+        $.ajax({
+            method: "PUT",
+            data: {
+                "description": description
+            },
+            url: baseUrl + "/task/" + $(this).data("taskid") + "/update"
+        }).done(function () {
+            //Success
+        }).error(function (response) {
+            console.error("Error:", response.responseText);
+        });
+    }
 });
